# openapi.yaml
openapi: 3.0.0
info:
  title: API GYM (Academia)
  description: API para gerenciamento de clientes e autenticação de administradores.
  version: 1.0.0
  contact:
    name: João Victor e Guilherme

servers:
  - url: http://localhost:5000
    description: Servidor Local de Desenvolvimento

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Insira o token JWT obtido na rota /login para acessar rotas privadas."

  schemas:
    Cliente:
      type: object
      properties:
        id:
          type: integer
          description: ID incremental gerado pelo sistema.
        cpf:
          type: string
          description: CPF do cliente (armazenado como string).
        nome:
          type: string
        status:
          type: boolean
          description: Indica se o cliente está ativo.
      example:
        id: 1
        cpf: "12345678900"
        nome: "João Silva"
        status: true

    Erro:
      type: object
      properties:
        error:
          type: string

paths:
  /:
    get:
      summary: Informações iniciais da API
      responses:
        '200':
          description: Retorna o nome da API, versão e autores.
          content:
            application/json:
              schema:
                type: object
                properties:
                  api: {type: string}
                  version: {type: string}
                  authors: {type: string}

  /clientes:
    get:
      summary: Lista todos os clientes cadastrados
      responses:
        '200':
          description: Lista de objetos de clientes.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cliente'

  /cliente:
    post:
      summary: Cadastra um novo cliente
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cpf, nome]
              properties:
                cpf: {type: string}
                nome: {type: string}
      responses:
        '201':
          description: Usuário cadastrado com sucesso.
        '400':
          description: Dados inválidos ou falha no cadastro.

  /cliente/{cpf}:
    parameters:
      - name: cpf
        in: path
        required: true
        description: CPF do cliente (numérico na URL).
        schema:
          type: integer
    get:
      summary: Busca um cliente pelo CPF
      responses:
        '200':
          description: Dados do cliente encontrados.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cliente'
        '404':
          description: Nenhum cliente encontrado com este CPF.

    patch:
      summary: Altera informações parciais do cliente
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nome: {type: string}
                cpf: {type: string}
                status: {type: boolean}
      responses:
        '200':
          description: Cliente alterado com sucesso.
        '400':
          description: Dados inválidos.
        '404':
          description: Cliente não encontrado.

    delete:
      summary: Exclui um cliente pelo CPF
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Cliente excluído com sucesso.
        '404':
          description: Cliente não encontrado.
        '500':
          description: Erro interno no processamento.

  /login:
    post:
      summary: Realiza login para obter o token de administrador
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [usuario, senha]
              properties:
                usuario: {type: string}
                senha: {type: string}
      responses:
        '200':
          description: Login realizado com sucesso.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: {type: string}
                  token: {type: string}
        '400':
          description: Dados de login não enviados.
        '401':
          description: Usuário ou senha inválidos.